NuGet packaging: refactored NuGetPackage and StaticPackage to their own two files
